Explicitly request Python 2
Our version of Mercurial requires python version 2 to run. Starting
with Ubuntu 20.04 the installed python defaults to version 3. This
change explicitly calls python2 so that we run with the correct
version. `/usr/bin/python2` is available at least since Ubuntu 18.04,
so this shouldn't cause problems on the currently supported systems.