Allow extraction to a custom location
WeSay's solution is in the src directory, which is a sibling
of the Output directory. Chorus, when searching for Mercurial,
will look in parent directories of the executing assembly's
directory. Since WeSay's SolutionDir is not a parent of the
output dir, Mercurial won't be found. This will allow WeSay
and similar projects to specify a custom location that will
be found.