Update Mercurial to 6.5.1 for Windows and Linux (#11)
* Update fixutf8.py to be Python 3 compatible

* Deal with changes from Mercurial 3.3 to 6.5.1

* All tests working on Windows

Green board! And that's even after getting rid of wrapping
Mercurial's system() function, so the win32helper.py file
can be essentially eliminated now.

Note that some Chorus changes do need to happen at the same time
for this to work: specifically, ChorusMerge's Program.cs does some
encoding tricks to decode cp1252 and reencode in UTF-8, and that's
no longer needed now.

* Remove leftover linux-x86 files

* Update linux-x64 mercurial to 6.5.1

This copy of Mercurial comes from the source download at
https://www.mercurial-scm.org/downloads

* Update Windows Mercurial

These files come from the TortoiseHG installation MSI available from
https://www.mercurial-scm.org/downloads, stripped of most of the files
relating to TortoiseHG as we'll never need to run thg.exe as a Windows
application, only hg.exe as a command-line app.

* Bump package version to 6.5.1

* Add defaultrc to windows Mercurial package

Apparently needed by TortoiseHG 6.5.1's default hg.exe config.

* Improve fixutf8 extension, including one bugfix

There was one important bugfix commented out by accident, leading to at
least one "repo is not empty" error when the repo was, in fact, empty.
That is fixed in this commit, plus I've rearranged the code so it's a
little cleaner and easier to read.

* Get rid of Qt DLLs in NuGet package

The Qt DLLs (and .pyd files and so on) from TortoiseHg are not needed
in Chorus, so let's trim the .nupkg size by removing them.

* Remove commented-out code accidentally left in

I missed this section of commented-out code earlier when cleaning up.