Add .pyc files for fixutf8, and add more guidsForInstaller.xml files (#13)
* Add .pyc files for fixutf8 extension

This is mostly for the sake of Windows machines, where the fixutf8
extension is likely to end up installed somewhere like Program Files
where non-admin accounts don't have write rights. Since the Windows
version of Mercurial that we distribute comes with Python 3.9, we'll
distribute .pyc files compiled with Python 3.9 as well.

* Add more .guidsForInstaller.xml files

Mercurial 6.5.1 comes with several more folders that need to be in the
NuGet package, and each folder needs its own guidsForInstaller.xml file.