Add build step to build chg binary (#25)
* Add build step to build chg binary

This is much faster when starting hg repeatedly for short commands,
which is what Chorus does a lot.

* Build Python 3.12 version with Python 3.12.4

Python 3.12.5 and later produce an error when building this specific
version of Mercurial, probably due to a name conflict. Rather than track
it down, we can safely build with 3.12.4.

* Build Python 3.13 version with Python 3.13.0-beta.3

Python 3.13.0-beta.4 and later produce an error when building this
specific version of Mercurial, probably the same one that Python 3.12.5
fails with. It should be safe to build Python 3.13 versions with beta.3.

* Remove executable bit from files that don't need it

Only the hg and chg binaries should actually be executable.

* Add results of building chg

The various .so files have changed as well because they were last build
six months ago, with different Python patch releases. But the source is
the same as the build from six months ago, so no .py files have changed.

* Update NuGet CI workflow to build Linux hg in CI

Instead of running the "Build Mercurial" workflow manually, we will now
run its process on every PR merged to master. That will cause less
confusion in the GHA actions log.

* Remove separate "Build Mercurial" workflow

Now that it's been incorporated into the nuget-ci-cd.yml workflow,
there's no more need for the build-hg.yml to be a separate workflow.
Having just one single workflow makes the GitHub Actions logs eaiser
to understand.