Switch NuGet build to use `dotnet pack`
This also allows for using the `buildTransitive` directory so that
packages will no longer need to consume the SIL.Chorus.Mercurial package
directly in order to set Mercurial4ChorusDestDir and have Mercurial
files copied to the right place. Now it will be sufficient to consume
the Chorus NuGet package, and the targets from SIL.Chorus.Mercurial will
automatically be used to copy the Mercurial files into position.